SafeScan/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
httpx==0.27.2
google-genai==1.14.0
python-dotenv==1.0.1
